Remove old ExecuteTaskBuildOperationType
